+++ src/main/java/com/molybdenum/alloyed/common/registry/ModCompatBlocks.java
@@ -11,7 +11,7 @@ import com.molybdenum.alloyed.common.item.ModItemGroup;
 import com.molybdenum.alloyed.data.util.BlockStateUtils;
 import com.molybdenum.alloyed.data.util.ModelUtils;
 import com.simibubi.create.foundation.data.CreateRegistrate;
-import com.simibubi.create.repack.registrate.util.entry.BlockEntry;
+import com.tterrag.registrate.util.entry.BlockEntry;
 import net.minecraft.advancements.critereon.StatePropertiesPredicate;
 import net.minecraft.client.renderer.RenderType;
 import net.minecraft.tags.BlockTags;