+++ src/main/java/com/molybdenum/alloyed/data/advancements/DisplayInfoBuilder.java
@@ -2,7 +2,7 @@ package com.molybdenum.alloyed.data.advancements;
 
 import com.molybdenum.alloyed.Alloyed;
 import com.molybdenum.alloyed.data.providers.ModAdvancementProvider;
-import com.simibubi.create.repack.registrate.providers.ProviderType;
+import com.tterrag.registrate.providers.ProviderType;
 import net.minecraft.advancements.DisplayInfo;
 import net.minecraft.advancements.FrameType;
 import net.minecraft.network.chat.TranslatableComponent;