+++ src/main/java/com/molybdenum/alloyed/data/util/DataUtils.java
@@ -1,8 +1,7 @@
 package com.molybdenum.alloyed.data.util;
 
-import com.simibubi.create.repack.registrate.builders.BlockBuilder;
-import com.simibubi.create.repack.registrate.util.nullness.NonNullFunction;
-import net.minecraft.tags.Tag;
+import com.tterrag.registrate.builders.BlockBuilder;
+import com.tterrag.registrate.util.nullness.NonNullFunction;
 import net.minecraft.tags.TagKey;
 import net.minecraft.world.item.Item;
 import net.minecraft.world.level.block.Block;