+++ changelog.md
@@ -1,4 +1,8 @@
-## Create: Alloyed v1.4b <br/> 
+## Create: Alloyed v1.4c <br/> 
 ### Fixed: <br/> 
-- Mod not working on Create 0.5b <br/> 
+- Mod not working on Create 0.5.0d <br/> 
+- Advancements not working <br/> 
+### Changed: <br/> 
+- Steel Shovel texture <br/> 
+- Steel Tools to be at iron mining level <br/> 
 #### If you find an issue with the mod, please submit it&nbsp;<a href="https://github.com/MythrilBagels/Create-Alloyed/issues" rel="nofollow">here</a>.